[move/#34] JopStepModel -> date 디렉토리 이동
