+++ app/src/main/java/com/haphap/app/data/model/detail/JobStepModel.kt
@@ -1,8 +1,8 @@
-package com.haphap.app.presentation.jobdetail.model
+package com.haphap.app.data.model.detail
 
 import com.haphap.app.presentation.jobdetail.type.JobStepStatus
 
-data class JobStep(
+data class JobStepModel(
     val number: Int,
     val stageName: String,
     val status: JobStepStatus,
